Moved providers to config
